requirements.txt
boto3
langchain
langchain_community
chromadb
sentence-transformers
requests
streamlit
transformers
torch
langchain-huggingface
langchain-chroma
langchain-aws
rank_bm25 
faiss-cpu
python-dotenv
google-genai
fastapi
uvicorn
langgraph
langgraph-checkpoint-redis
langgraph-checkpoint-aws
langgraph-checkpoint-aws[valkey]
google-genai
langchain-text-splitters
langchain-redis